requirements.txt
aws-cdk-lib>=2.131.0
constructs>=10.3.0
aws-cdk.aws-lambda-python-alpha>=2.131.0.a0
boto3>=1.34.0
pydantic>=2.0.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.2.0
requests>=2.31.0
pytz>=2024.1
ulid-py>=1.1.0
timezonefinder>=6.2.0

